# gen/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- description: Generated server url
  url: http://localhost:8080
paths:
  /vehicles/{id}:
    delete:
      operationId: deleteVehicle
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - vehicle-controller
      x-accepts: application/json
    get:
      operationId: getVehicleById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - vehicle-controller
      x-accepts: '*/*'
    put:
      operationId: updateVehicle
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - vehicle-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /user-page/vehicles/{id}:
    delete:
      operationId: deleteUserVehicle
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - user-page-controller
      x-accepts: application/json
    get:
      operationId: getUserVehicleById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - user-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateUserVehicle
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - user-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /admin-page/vehicles/{id}:
    delete:
      operationId: deleteVehicle_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - admin-page-controller
      x-accepts: application/json
    get:
      operationId: getVehicleById_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    put:
      operationId: updateVehicle_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /vehicles:
    get:
      operationId: getVehicles
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehicleDto'
                type: array
          description: OK
      tags:
      - vehicle-controller
      x-accepts: '*/*'
    post:
      operationId: createVehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - vehicle-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /vehicle-photos:
    get:
      operationId: getAllPhotos
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehiclePhotoDto'
                type: array
          description: OK
      tags:
      - vehicle-photo-controller
      x-accepts: '*/*'
    post:
      operationId: addPhotoToVehicle
      parameters:
      - explode: true
        in: query
        name: vehicleId
        required: true
        schema:
          format: int64
          type: integer
        style: form
      - explode: true
        in: query
        name: isMain
        required: true
        schema:
          type: boolean
        style: form
      - explode: true
        in: query
        name: username
        required: true
        schema:
          type: string
        style: form
      requestBody:
        $ref: '#/components/requestBodies/inline_object'
        content:
          application/json:
            schema:
              properties:
                photo:
                  format: binary
                  type: string
              required:
              - photo
              type: object
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehiclePhoto'
          description: OK
      tags:
      - vehicle-photo-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /admin-page/vehicles:
    get:
      operationId: getAllVehicles
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehicleDto'
                type: array
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
    post:
      operationId: createVehicle_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleDto'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehicleDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
  /admin-page/vehicle-photos/{vehicleId}:
    post:
      operationId: addPhotoToVehicle_1
      parameters:
      - explode: false
        in: path
        name: vehicleId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: true
        in: query
        name: isMain
        required: true
        schema:
          type: boolean
        style: form
      requestBody:
        $ref: '#/components/requestBodies/inline_object_2'
        content:
          application/json:
            schema:
              properties:
                photo:
                  format: binary
                  type: string
              required:
              - photo
              type: object
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehiclePhotoDto'
          description: OK
      tags:
      - admin-page-controller
      x-contentType: application/json
      x-accepts: '*/*'
...
  /vehicle-photos/{id}:
    delete:
      operationId: deletePhoto
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - vehicle-photo-controller
      x-accepts: application/json
    get:
      operationId: getPhotoById
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehiclePhotoDto'
          description: OK
      tags:
      - vehicle-photo-controller
      x-accepts: '*/*'
...
  /admin-page/vehicles/user/{userId}:
    get:
      operationId: getVehiclesByUser
      parameters:
      - explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/VehicleDto'
                type: array
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
...
  /admin-page/vehicle-photos/{id}:
    delete:
      operationId: deletePhoto_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - admin-page-controller
      x-accepts: application/json
    get:
      operationId: getPhotoById_1
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VehiclePhotoDto'
          description: OK
      tags:
      - admin-page-controller
      x-accepts: '*/*'
...
  /user-page/photos/{photoId}:
    delete:
      operationId: deleteVehiclePhoto
      parameters:
      - explode: false
        in: path
        name: photoId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: OK
      tags:
      - user-page-controller
      x-accepts: application/json
components:
  requestBodies:
...
    inline_object:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object'
    inline_object_2:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_2'
  schemas:
    VehicleDto:
      example:
        color: color
        year: 6
        description: description
        seats: 2
        hasWarranty: true
        advertisementId: 7
        condition: NEW
        transmission: MANUAL
        fuelType: PETROL
        price: 1.4658129805029452
        model: model
        vin: vin
        id: 0
        power: 0
        brand: brand
        mileage: 5
      properties:
        id:
          format: int64
          type: integer
        brand:
          maxLength: 50
          minLength: 1
          type: string
        model:
          maxLength: 50
          minLength: 1
          type: string
        year:
          format: int32
          type: integer
        price:
          format: double
          type: number
        fuelType:
          enum:
          - PETROL
          - DIESEL
          - LPG
          - ELECTRIC
          - HYBRID
          type: string
        transmission:
          enum:
          - MANUAL
          - AUTOMATIC
          type: string
        color:
          maxLength: 20
          minLength: 1
          type: string
        mileage:
          format: int32
          type: integer
        power:
          format: int32
          minimum: 0
          type: integer
        seats:
          format: int32
          maximum: 8
          minimum: 1
          type: integer
        description:
          maxLength: 2000
          minLength: 0
          type: string
        vin:
          maxLength: 17
          minLength: 17
          type: string
        condition:
          enum:
          - NEW
          - USED
          - DAMAGED
          type: string
        hasWarranty:
          type: boolean
        advertisementId:
          format: int64
          type: integer
      required:
      - advertisementId
      - brand
      - color
      - condition
      - fuelType
      - hasWarranty
      - model
      - price
      - transmission
      - vin
      - year
      type: object
...
    Advertisement:
      example:
        id: 0
        title: title
        user:
          favorites:
          - 1
          - 1
          advertisements:
          - null
          - null
          password: password
          role: ADMIN
          credentialsNonExpired: true
          accountNonExpired: true
          id: 6
          email: email
          enabled: true
          authorities:
          - authority: authority
          - authority: authority
          username: username
          accountNonLocked: true
      properties:
        id:
          format: int64
          type: integer
        title:
          type: string
        user:
          $ref: '#/components/schemas/User'
      type: object
    GrantedAuthority:
      example:
        authority: authority
      properties:
        authority:
          type: string
      type: object
    User:
      example:
        favorites:
        - 1
        - 1
        advertisements:
        - null
        - null
        password: password
        role: ADMIN
        credentialsNonExpired: true
        accountNonExpired: true
        id: 6
        email: email
        enabled: true
        authorities:
        - authority: authority
        - authority: authority
        username: username
        accountNonLocked: true
      properties:
        id:
          format: int64
          type: integer
        advertisements:
          items:
            $ref: '#/components/schemas/Advertisement'
          type: array
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          enum:
          - ADMIN
          - USER
          - GUEST
          type: string
        enabled:
          type: boolean
        favorites:
          items:
            format: int64
            type: integer
          type: array
        authorities:
          items:
            $ref: '#/components/schemas/GrantedAuthority'
          type: array
        accountNonExpired:
          type: boolean
        accountNonLocked:
          type: boolean
        credentialsNonExpired:
          type: boolean
      type: object
    Vehicle:
      example:
        color: color
        year: 1
        description: description
        seats: 7
        hasWarranty: true
        condition: NEW
        transmission: MANUAL
        fuelType: PETROL
        price: 5.962133916683182
        advertisement:
          id: 0
          title: title
          user:
            favorites:
            - 1
            - 1
            advertisements:
            - null
            - null
            password: password
            role: ADMIN
            credentialsNonExpired: true
            accountNonExpired: true
            id: 6
            email: email
            enabled: true
            authorities:
            - authority: authority
            - authority: authority
            username: username
            accountNonLocked: true
        model: model
        vin: vin
        id: 6
        power: 2
        brand: brand
        mileage: 5
      properties:
        id:
          format: int64
          type: integer
        brand:
          type: string
        model:
          type: string
        year:
          format: int32
          type: integer
        price:
          format: double
          type: number
        fuelType:
          enum:
          - PETROL
          - DIESEL
          - LPG
          - ELECTRIC
          - HYBRID
          type: string
        transmission:
          enum:
          - MANUAL
          - AUTOMATIC
          type: string
        color:
          type: string
        mileage:
          format: int32
          type: integer
        power:
          format: int32
          type: integer
        seats:
          format: int32
          type: integer
        description:
          type: string
        vin:
          type: string
        condition:
          enum:
          - NEW
          - USED
          - DAMAGED
          type: string
        hasWarranty:
          type: boolean
        advertisement:
          $ref: '#/components/schemas/Advertisement'
      type: object
    VehiclePhoto:
      example:
        photoUrl: photoUrl
        isMain: true
        id: 0
        vehicle:
          color: color
          year: 1
          description: description
          seats: 7
          hasWarranty: true
          condition: NEW
          transmission: MANUAL
          fuelType: PETROL
          price: 5.962133916683182
          advertisement:
            id: 0
            title: title
            user:
              favorites:
              - 1
              - 1
              advertisements:
              - null
              - null
              password: password
              role: ADMIN
              credentialsNonExpired: true
              accountNonExpired: true
              id: 6
              email: email
              enabled: true
              authorities:
              - authority: authority
              - authority: authority
              username: username
              accountNonLocked: true
          model: model
          vin: vin
          id: 6
          power: 2
          brand: brand
          mileage: 5
      properties:
        id:
          format: int64
          type: integer
        photoUrl:
          type: string
        isMain:
          type: boolean
        vehicle:
          $ref: '#/components/schemas/Vehicle'
      type: object
    VehiclePhotoDto:
      example:
        photoUrl: photoUrl
        id: 0
        vehicleId: 6
      properties:
        id:
          format: int64
          type: integer
        photoUrl:
          type: string
        vehicleId:
          format: int64
          type: integer
      type: object
...
    inline_object:
      properties:
        photo:
          format: binary
          type: string
      required:
      - photo
      type: object
...
    inline_object_2:
      properties:
        photo:
          format: binary
          type: string
      required:
      - photo
      type: object